Update patch set 2
Patch Set 2:

As a general note the aaiContextBuilder URI is:
/aaicontextbuilder/service/context
so it was never versioned.
Sharon wants it versioned but I don't know if it should be v1 or v2 at this point.
We should clarify this and fix it as part of this update

Patch-set: 2
Reviewer: Gerrit User 3131 <3131@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0